+++ python/dalex/dalex/__init__.py
@@ -1,7 +1,7 @@
 # dx.Explainer
 from ._explainer.object import Explainer
 
-# from dalex.datasets import load_titanic
+# from dalex.datasets import load_*
 from . import datasets